Updated README to reflect change in system design, renamed baseWeight to baseDeliveryCost, added costmanager and unit tests, changed all Offer numbers to Double

+++ EverestPackageDeliveryCLI/Models/Offer.swift
@@ -9,9 +9,9 @@ import Foundation
 
 public struct Offer {
   let offerID: String
-  let lowerBoundWeightInKg: Int
-  let upperBoundWeightInKg: Int
-  let lowerBoundDistanceInKm: Int
-  let upperBoundDistanceInKm: Int
+  let lowerBoundWeightInKg: Double
+  let upperBoundWeightInKg: Double
+  let lowerBoundDistanceInKm: Double
+  let upperBoundDistanceInKm: Double
   let discountRateInPercent: Double
 }
